+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -86,7 +86,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JVY60BJW3XHZ3YEQJWM2944T"
+        "01JVY7BNPKXG5CWS11VVGZXP67"
     )
     r"""The id of the resource"""