+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -100,7 +100,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JVY60BK4XGC61KDYCDGENQEB"
+        "01JVY7BNPNY3WAFMVBW3ZHHFBC"
     )
     r"""The id of the resource"""