+++ packages/orq-rc/RELEASES.md
@@ -1449,3 +1449,13 @@ Based on:
 - [python v3.6.0-rc.30] packages/orq-rc
 ### Releases
 - [PyPI v3.6.0-rc.30] https://pypi.org/project/orq-ai-sdk/3.6.0-rc.30 - packages/orq-rc
+
+## 2025-05-23 09:26:23
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.550.0 (2.610.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.6.0-rc.31] packages/orq-rc
+### Releases
+- [PyPI v3.6.0-rc.31] https://pypi.org/project/orq-ai-sdk/3.6.0-rc.31 - packages/orq-rc